refactor(tests): extract shared admin/regular user and client fixtures
The admin/regular user and client fixtures were duplicated verbatim (modulo
usernames) across db and forwarding view tests. Promote them to
tests/fixtures.py and convert apps/db/tests/test_views.py to import them.
